backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
passlib[bcrypt]>=1.7.4

# Testing dependencies
pytest==7.4.3
pytest-bdd==7.0.1
httpx==0.25.2
pytest-cov==4.1.0
pytest-asyncio==0.21.1
flake8>=6.0.0
mypy>=1.0.0
apscheduler
PyJWT>=2.8.0
